Move ping and traceroute to Netceptor package
